Part, Testcase, Unitの型定義を追加
